backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
sendgrid>=6.12.0
sentence-transformers==3.1.0
torch>=2.0.0
transformers>=4.33.0
email-validator==2.2.0
